resolved temp.. and humidity data precision issue, timzone of sensor and also documented the sensor generatoin script

+++ init-scripts/init.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS sensor_data (
     id SERIAL PRIMARY KEY,
     sensor_id INT NOT NULL,
-    temperature FLOAT NOT NULL,
-    humidity FLOAT NOT NULL,
+    temperature NUMERIC(5,2) NOT NULL, 
+    humidity NUMERIC(5,2) NOT NULL,     
     timestamp TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
